fix: remove silly foreground/background concept to fix fps issue
